inventory-abstraction/requirements.txt

ultralytics
norfair
opencv-python
pyzbar
pandas
Pillow
imgaug
numpy
